refactor: move symbol table to separate files
